:bear: moved App into it's on ES6 component
